+++ CHANGELOG.md
@@ -10,6 +10,7 @@ with regard to its JavaScript and TypeScript bindings' APIs and the Ethereum ABI
 ### Added
 - Burning capability to Radicle Token
 - Radicle Link ID claims contract
+- Pool events when a direct stream from a sender to a receiver is updated
 
 ### Changed
 - Replaced vesting contract with one adapted from Melonport